move predict to trainer
